Remove módulos agregados antigos do domínio
