+++ sentinela/domain/ports/__init__.py
@@ -1,0 +1,5 @@
+"""Portas que conectam o domínio com outros serviços e adaptadores."""
+from .article_sink import ArticleSink
+from .portal_gateway import PortalGateway
+
+__all__ = ["PortalGateway", "ArticleSink"]
