+++ sentinela/services/publications/api.py
@@ -11,7 +11,7 @@ from fastapi import APIRouter, FastAPI, HTTPException
 from fastapi.middleware.cors import CORSMiddleware
 from pydantic import BaseModel
 
-from sentinela.domain.entities import Article
+from sentinela.domain import Article
 from sentinela.services.publications import (
     PublicationsContainer,
     build_publications_container,
